tests: Complete code coverage to 100%
Add missing test case for invalid input to getRelativePath().
This is the same test case we use for joinPath() as well.

Also ignore coverage for default value of $start. This is already
covered by tests, but by mocking the start directory. There is no
need to cover the actual call to PHP's getcwd().

Change-Id: I40e62f587df4fb22745bb62c8cf31f3128d2d7d0