Drop support for PHP 5.3 and PHP 5.4
Per T75901, MediaWiki no longer supports this. Our CI environment
also hasn't been verifying this library on PHP 5.3 or PHP 5.4 for
a while.

Travis CI also hasn't supported PHP 5.3 for a while. PHP 5.4 still
works, but if we're bumping it now, we might as well go straight
to the current minimum requirement so that we can sustain this
level for a bit longer without requiring another major version
bump.

Change-Id: Iffb7b707e2ef04895db3d18ed2db8494a99883cf