RelPath: Add strict type hints to joinPath() and other RelPath methods
We have at least one report of some bad code or cache corruption
causing false to be passed in:

```
@1 vendor/…/RelPath.php: Wikimedia\RelPath::splitPath(bool)
@2 mediawiki/…/ResourceLoaderModule.php: Wikimedia\RelPath::joinPath(str, bool)
```

This led to unsupported input with false becoming an empty string, which then
caused an Undefined string offset warning, with further unspecified and
non-sensical return values. This change turns any such error into a fatal,
to be mitigated by fixing the caller higher up in the stack (if it happens again).

Bug: T264203
Change-Id: I7f37371db00fcf529e7bc501ca4762f10c991be5